Added got_hit and game_over method for player class
This method is meant to manage the lives of the player. The player can only have 6 lives which get less as the player is hit by the invader's lasers, this is achieved in the got_hit method. I also implemented a game_over method that will check if the player still has a life, if not it will change a value in the Environment class that will stop the game from running.