Adds sprite graphics and changes draw/move methods
This commit adds sprites for objects that appears when the
player launches the game. The screen contains black background,
an image of a player, and images of all enemies. Methods draw annd move
relate to pygame.sprite rather than to pygame geometric drawings.

This class is essential in order to visualize all gameplay.